clientのPlayerServiceをRunnableに修正
