+++ client/src/main/java/com/example/client/PlayerSyncResponseObserver.java
@@ -4,10 +4,7 @@ import com.example.shared.GrpcLocation;
 import com.example.shared.GrpcPlayer;
 import com.example.shared.PlayerSyncRequest;
 import com.example.shared.PlayerSyncResponse;
-import com.google.common.util.concurrent.Uninterruptibles;
 import io.grpc.stub.StreamObserver;
-
-import java.util.concurrent.TimeUnit;
 
 public class PlayerSyncResponseObserver implements StreamObserver<PlayerSyncResponse> {
 
@@ -24,11 +21,8 @@ public class PlayerSyncResponseObserver implements StreamObserver<PlayerSyncResp
   @Override
   public void onNext(PlayerSyncResponse value) {
     GrpcPlayer otherGrpcPlayer = value.getOtherPlayer();
-    if (!player.id().equals(otherGrpcPlayer.getId())) {
-      OtherPlayer otherPlayer = convert(otherGrpcPlayer);
-      otherPlayers.moveOrAdd(otherPlayer);
-    }
-    syncPlayer();
+    OtherPlayer otherPlayer = convert(otherGrpcPlayer);
+    otherPlayers.moveOrAdd(otherPlayer);
   }
 
   @Override
@@ -46,9 +40,7 @@ public class PlayerSyncResponseObserver implements StreamObserver<PlayerSyncResp
     syncPlayer();
   }
 
-  private void syncPlayer() {
-    Uninterruptibles.sleepUninterruptibly(32, TimeUnit.MILLISECONDS);
-
+  public void syncPlayer() {
     Location location = player.location();
     GrpcLocation grpcLocation = GrpcLocation.newBuilder().setX(location.getX()).setY(location.getY()).build();
     GrpcPlayer grpcPlayer = GrpcPlayer.newBuilder().setId(player.id()).setName(player.name()).setLocation(grpcLocation).build();
